feat: add workflow extraction utilities for ComfyUI Prompt Manager
- Implemented shared extraction helpers for sampler parameters, VAE info, CLIP info, resolution, and model resolution.
- Added functions to extract sampler parameters, VAE loader information, CLIP loader information, and image/video resolution.
- Introduced model name resolution functions for models, VAEs, and CLIPs.
- Created a comprehensive extraction function to gather all generation parameters from various file types.
- Included a function to enrich extraction results with availability flags for LoRAs, models, VAE, and CLIP.
- Built a simplified workflow data structure for use in WorkflowGenerator and PromptExtractor nodes.
