+++ py/workflow_node_utils.py
@@ -1,6 +1,6 @@
 """
 Thin re-exports of node-graph helpers from prompt_extractor.
-Placed in py/ so workflow_extractor_utils can import without a circular dep.
+Placed in py/ so workflow_extraction_utils can import without a circular dep.
 Uses relative import since both py/ and nodes/ are siblings under the package root.
 """
 from ..nodes.prompt_extractor import build_node_map, build_link_map
